deploy/requirements.txt
GitPython
datadog
slacker